Merchant feature

+++ src/main/java/com/JustHealth/Health/DTO/InventoryDTO.java
@@ -1,5 +1,6 @@
 package com.JustHealth.Health.DTO;
 
+import com.JustHealth.Health.Entity.Product;
 import lombok.Data;
 
 import java.util.Date;
@@ -16,7 +17,6 @@ public class InventoryDTO {
     private Integer minQTY;
     private Integer maxQTY;
     private Integer GST;
-
-    private Integer productId;
+    private Long productId;
 
 }
